+++ src/features/common/routes/Error/404/style.ts
