+++ src/features/common/routes/HelpDesk/Success/style.ts
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+import { SelectContainer as BaseContainer } from "../Select/style";
+
+export const SuccessContainer = styled(BaseContainer)`
+  width: 100%;
+`;
